[ADD] Inventory Module Infrastructure

+++ dhHelpdeskNG/dhHelpdesk_NG.Domain/WorkstationModules/LogicalDrive.cs
@@ -6,8 +6,8 @@
 
     public class LogicalDrive : Entity
     {
-        public double FreeBytes { get; set; }
-        public double TotalBytes { get; set; }
+        public decimal FreeBytes { get; set; }
+        public decimal TotalBytes { get; set; }
         public int Computer_Id { get; set; }
         public int DriveType { get; set; }
         public string DriveLetter { get; set; }
